[FLINK-19070][hive] Hive connector should throw a meaningful exception if user reads/writes ACID tables
