--- flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/connectors/hive/HiveUtils.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.connectors.hive;
-
-import org.apache.flink.table.catalog.CatalogTable;
-import org.apache.flink.table.catalog.ObjectPath;
-
-/**
- * Utils for Hive connector.
- */
-public class HiveUtils {
-
-	/**
-	 * Check whether to read or write on the hive ACID table.
-	 *
-	 * @param catalogTable Hive catalog table.
-	 * @param tablePath    Identifier table path.
-	 * @throws FlinkHiveException Thrown, if the source or sink table is transactional.
-	 */
-	public static void checkAcidTable(CatalogTable catalogTable, ObjectPath tablePath) {
-		if (catalogTable != null && catalogTable.getOptions() != null) {
-			String tableIsTransactional = catalogTable.getOptions().get("transactional");
-			if (tableIsTransactional == null) {
-				tableIsTransactional = catalogTable.getOptions().get("transactional".toUpperCase());
-			}
-			if (tableIsTransactional != null && tableIsTransactional.equalsIgnoreCase("true")) {
-				throw new FlinkHiveException(String.format("Reading or writing ACID table %s is not supported.", tablePath));
-			}
-		}
-	}
-}
